Rename live as digest (as before)
Separate digest-specific actions.
